added some doc added license info added version info

+++ src/jsingleinstace/JSingleInstance.java
@@ -1,3 +1,22 @@
+/**
+ * JSingleInstance - allows running only one instance of any java app
+ * Copyright (C) 2011 MJ <mj_dv@web.de>
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+
 package jsingleinstace;
 
 import java.io.BufferedReader;
@@ -17,12 +36,40 @@ import java.util.EventObject;
 import java.util.Iterator;
 import java.util.List;
 
+/**
+ * 
+ * @author MJ
+ * a little class to allow only one instance of an application
+ * {@link #JSingleInstance(String)} constructs a new object
+ * {@link #isAlreadyRunning()} checks if an instance is already running
+ * {@link #sendCommand(String)} send a command to the running instance
+ * to receive commands just add an CommandListener
+ * {@link #exit()} exits the instance, must be called when you exit your app!
+ * but not if there is already an instance running
+ * 
+ */
 public class JSingleInstance {
 	
+	/**
+	 * represents the version string
+	 */
+	public final static String VERSION = "0.1";
+	
+	/**
+	 * @author MJ
+	 * interface to receive commands
+	 *
+	 */
 	public interface CommandListener {
 		public void onCommand(CommandEvent e);
 	}
 	
+	/**
+	 * Represents a command event
+	 * {@link CommandEvent#command} represents the command which has been sent
+	 * @author MJ
+	 *
+	 */
 	public class CommandEvent extends EventObject {
 		/**
 		 * 
@@ -35,7 +82,7 @@ public class JSingleInstance {
 		}
 	}
 	
-	List<CommandListener> commandListeners = new ArrayList<CommandListener>();
+	private List<CommandListener> commandListeners = new ArrayList<CommandListener>();
 	
 	
 	private File f;
@@ -46,6 +93,13 @@ public class JSingleInstance {
 	private ServerSocket serverSocket;
 	private PrintWriter clientOut;
 	
+	/**
+	 * constructs a new object
+	 * @param path file where instance info (like port and pid) can be stored
+	 * should be something like "USERFILES\myapp.info"
+	 * @throws IOException thrown if path is not writeable or other instance
+	 * has not removed the info file (not called {@link #exit()}, or crashed)
+	 */
 	public JSingleInstance(String path) throws IOException {
 		this.f = new File(path);
 		
@@ -112,7 +166,12 @@ public class JSingleInstance {
 			
 		}).start();
 	}
-	
+	/**
+	 * sends a command to existing instance
+	 * @param cmd the command which should be send
+	 * must not contain new line ('\n') because it is used to seperate commands!
+	 * @return success or not
+	 */
 	public boolean sendCommand(String cmd) {
 		if(!isAlreadyRunning) return false;
 		
@@ -120,11 +179,22 @@ public class JSingleInstance {
 		return true; //TODO wait for ok
 	}
 	
+	/**
+	 * exits the instance
+	 * has to be called to delete the info file and close the socket
+	 * DO NOT call if there is another instance running!
+	 * @throws IOException thrown if info file does not exist
+	 */
 	public void exit() throws IOException {
+		if(isAlreadyRunning) return;
 		serverSocket.close();
 		f.delete();
 	}
 	
+	/**
+	 * checks if another instance is already running
+	 * @return true if there is another instance running
+	 */
 	public boolean isAlreadyRunning() {
 		return isAlreadyRunning;
 	}
